% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_tables.R
\name{apa_descriptives_table}
\alias{apa_descriptives_table}
\title{Create APA-style descriptives table}
\usage{
apa_descriptives_table(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variable names}
}
\value{
HTML table
}
\description{
Create APA-style descriptives table
}
