% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_tables.R
\name{apa_table_wrapper}
\alias{apa_table_wrapper}
\title{Generate APA-style HTML table wrapper with copy button}
\usage{
apa_table_wrapper(table_html, caption = "")
}
\arguments{
\item{table_html}{HTML table content}

\item{caption}{Table caption}
}
\value{
HTML string with copy button
}
\description{
Generate APA-style HTML table wrapper with copy button
}
