% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{convert_count}
\alias{convert_count}
\title{Convert COUNT command -> rowSums of logical checks}
\usage{
convert_count(parsed, sav_info)
}
\description{
Convert COUNT command -> rowSums of logical checks
}
