% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{convert_if}
\alias{convert_if}
\title{Convert IF command -> dplyr::mutate with if_else}
\usage{
convert_if(parsed, sav_info)
}
\arguments{
\item{parsed}{Parsed SPSS command object}

\item{sav_info}{SAV file information from parse_sav()}
}
\description{
Convert IF command -> dplyr::mutate with if_else
}
\keyword{internal}
