% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sav.R
\name{convert_labelled_to_factors}
\alias{convert_labelled_to_factors}
\title{Convert haven labelled variables to factors}
\usage{
convert_labelled_to_factors(data)
}
\arguments{
\item{data}{Data frame from haven::read_sav()}
}
\value{
Data frame with labelled variables converted to factors
}
\description{
Convert haven labelled variables to factors
}
