% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{convert_select_if}
\alias{convert_select_if}
\title{Convert SELECT IF -> dplyr::filter}
\usage{
convert_select_if(parsed, sav_info)
}
\arguments{
\item{parsed}{Parsed SPSS command object}

\item{sav_info}{SAV file information from parse_sav()}
}
\description{
Convert SELECT IF -> dplyr::filter
}
\keyword{internal}
