% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{expand_spss_to_ranges}
\alias{expand_spss_to_ranges}
\title{Convert SPSS expression to R expression}
\usage{
expand_spss_to_ranges(expr, sav_info)
}
\arguments{
\item{expr}{SPSS expression string}

\item{sav_info}{SAV file information from parse_sav()}
}
\value{
Expression with TO ranges expanded
}
\description{
Translates SPSS syntax expressions (operators, functions, variable names)
Expand SPSS TO ranges using variable order from SAV file
}
\details{
Converts "A11 TO A18" to "A11, A12, A13, A14, A15, A16, A17, A18"
by looking up variable positions in the SAV data dictionary.
}
\keyword{internal}
