% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{expand_to_syntax}
\alias{expand_to_syntax}
\title{Expand TO in a vector of variable names using a reference list}
\usage{
expand_to_syntax(vars, all_names)
}
\arguments{
\item{vars}{Character vector of variable names}

\item{all_names}{Character vector of all dataset variable names}
}
\description{
Expand TO in a vector of variable names using a reference list
}
\keyword{internal}
