% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{extract_command_type}
\alias{extract_command_type}
\title{Extract the main command type}
\usage{
extract_command_type(cmd)
}
\arguments{
\item{cmd}{Character string of SPSS command text}
}
\description{
Recognizes full and abbreviated SPSS command names.
}
\keyword{internal}
