% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{extract_options}
\alias{extract_options}
\title{Extract options from a command}
\usage{
extract_options(cmd, command_type)
}
\arguments{
\item{cmd}{Character string of SPSS command text}

\item{command_type}{Character string of SPSS command type}
}
\description{
Extract options from a command
}
\keyword{internal}
