% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{extract_pattern}
\alias{extract_pattern}
\title{Extract a pattern match from command}
\usage{
extract_pattern(cmd, pattern)
}
\arguments{
\item{cmd}{Character string of SPSS command text}

\item{pattern}{Regex pattern to extract}
}
\description{
Extract a pattern match from command
}
\keyword{internal}
