% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{extract_subcommands}
\alias{extract_subcommands}
\title{Extract subcommands from a command}
\usage{
extract_subcommands(cmd)
}
\arguments{
\item{cmd}{Character string of SPSS command text}
}
\description{
Subcommands start with / and contain key=value or just key
}
\keyword{internal}
