% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variables from a command}
\usage{
extract_variables(cmd, command_type)
}
\arguments{
\item{cmd}{Character string of SPSS command text}

\item{command_type}{Character string of SPSS command type}
}
\description{
Extract variables from a command
}
\keyword{internal}
