% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{extract_variables_clause}
\alias{extract_variables_clause}
\title{Extract the VARIABLES= clause}
\usage{
extract_variables_clause(cmd)
}
\arguments{
\item{cmd}{Character string of SPSS command text}
}
\description{
Extract the VARIABLES= clause
}
\keyword{internal}
