% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_tables.R
\name{format_effect_size}
\alias{format_effect_size}
\title{Format effect size for APA style}
\usage{
format_effect_size(es, ci_lower = NULL, ci_upper = NULL, es_name = "d")
}
\arguments{
\item{es}{Effect size value}

\item{ci_lower}{Lower CI bound}

\item{ci_upper}{Upper CI bound}

\item{es_name}{Name of effect size (e.g., "d", "eta-squared")}
}
\value{
Formatted string
}
\description{
Format effect size for APA style
}
