% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sav.R
\name{generate_data_load_code}
\alias{generate_data_load_code}
\title{Generate R code to recreate the data loading}
\usage{
generate_data_load_code(sav_info, new_path = "data.sav")
}
\arguments{
\item{sav_info}{Result from parse_sav()}

\item{new_path}{Path to use in generated code}
}
\value{
Character string of R code
}
\description{
Generate R code to recreate the data loading
}
\examples{
\dontrun{
sav_info <- parse_sav("data.sav")
cat(generate_data_load_code(sav_info, new_path = "data.sav"))
}
}
