% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator
Returns x if not NULL, otherwise y}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand value}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null-coalescing operator
Returns x if not NULL, otherwise y
}
\keyword{internal}
