% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Log message with timestamp}
\usage{
log_message(msg, level = "INFO")
}
\arguments{
\item{msg}{Message to log}

\item{level}{Log level (INFO, WARN, ERROR)}
}
\description{
Log message with timestamp
}
