% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{normalize_whitespace}
\alias{normalize_whitespace}
\title{Normalize whitespace in syntax}
\usage{
normalize_whitespace(syntax)
}
\arguments{
\item{syntax}{Character string of SPSS syntax text}
}
\description{
Normalize whitespace in syntax
}
\keyword{internal}
