% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sav.R
\name{parse_sav}
\alias{parse_sav}
\title{Parse an SPSS .SAV file}
\usage{
parse_sav(sav_path)
}
\arguments{
\item{sav_path}{Path to the .SAV file}
}
\value{
List containing data, metadata, and summary info
}
\description{
Reads an SPSS .sav data file and extracts the data, variable metadata,
value labels, and summary information.
}
\examples{
\dontrun{
# Parse an SPSS .sav data file
sav_info <- parse_sav("survey_data.sav")
sav_info$n_obs
sav_info$n_vars
head(sav_info$metadata)
}
}
