% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{parse_single_command}
\alias{parse_single_command}
\title{Parse a single SPSS command}
\usage{
parse_single_command(cmd)
}
\arguments{
\item{cmd}{Character string of SPSS command text}
}
\description{
Parse a single SPSS command
}
\keyword{internal}
