% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_plots.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Create interaction plot for factorial designs}
\usage{
plot_interaction(data, dv, factor1, factor2, ci = 0.95)
}
\arguments{
\item{data}{Data frame}

\item{dv}{Dependent variable name}

\item{factor1}{First factor (x-axis)}

\item{factor2}{Second factor (lines)}

\item{ci}{Confidence interval level}
}
\value{
ggplot object
}
\description{
Create interaction plot for factorial designs
}
