% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quote_var}
\alias{quote_var}
\title{Safe variable quoting for R code
Backticks variable names with special characters}
\usage{
quote_var(var)
}
\arguments{
\item{var}{Variable name}
}
\value{
Properly quoted variable name
}
\description{
Safe variable quoting for R code
Backticks variable names with special characters
}
