% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{remove_comments}
\alias{remove_comments}
\title{Remove SPSS comments from syntax}
\usage{
remove_comments(syntax)
}
\arguments{
\item{syntax}{Character string of SPSS syntax text}
}
\description{
Remove SPSS comments from syntax
}
\keyword{internal}
