% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_as_numeric}
\alias{safe_as_numeric}
\title{Convert factor to numeric if possible
Handles haven labelled variables}
\usage{
safe_as_numeric(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Numeric vector
}
\description{
Convert factor to numeric if possible
Handles haven labelled variables
}
