% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sps.R
\name{split_commands}
\alias{split_commands}
\title{Split syntax into individual commands}
\usage{
split_commands(syntax)
}
\description{
SPSS commands are terminated by a period at the end of a line.
We must not split on periods inside string literals or numbers.
}
