% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_plots.R
\name{theme_apa}
\alias{theme_apa}
\title{APA theme for ggplot2}
\usage{
theme_apa(base_size = 12, base_family = "serif")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
ggplot2 theme
}
\description{
APA theme for ggplot2
}
