% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_vars}
\alias{validate_vars}
\title{Validate that required columns exist in data}
\usage{
validate_vars(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Required variable names}
}
\value{
TRUE if all exist, error otherwise
}
\description{
Validate that required columns exist in data
}
