% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{convert_all_commands}
\alias{convert_all_commands}
\title{Convert all parsed commands to R code}
\usage{
convert_all_commands(parsed_commands, sav_info)
}
\arguments{
\item{parsed_commands}{List of parsed commands from \code{\link[=parse_sps]{parse_sps()}}}

\item{sav_info}{SAV file info from \code{\link[=parse_sav]{parse_sav()}}}
}
\value{
List of conversion results, each containing r_code, packages, and analysis_type
}
\description{
Takes a list of parsed SPSS commands and converts each to equivalent R code.
}
\examples{
\dontrun{
parsed   <- parse_sps("analysis.sps")
sav_info <- parse_sav("data.sav")
converted <- convert_all_commands(parsed, sav_info)
converted[[1]]$r_code
converted[[1]]$analysis_type
}
}
