% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{convert_spss_expression}
\alias{convert_spss_expression}
\title{Convert SPSS expression to R expression}
\usage{
convert_spss_expression(expr)
}
\arguments{
\item{expr}{SPSS expression string}
}
\value{
R expression string
}
\description{
Translates SPSS syntax expressions (operators, functions, variable names)
to equivalent R expressions.
}
\examples{
convert_spss_expression("MEAN(var1, var2, var3)")
convert_spss_expression("MISSING(age)")
convert_spss_expression("x = 1 AND y <> 0")
}
