% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{expand_spss_variables}
\alias{expand_spss_variables}
\title{Expand SPSS TO syntax in variable lists}
\usage{
expand_spss_variables(parsed_command, sav_info)
}
\arguments{
\item{parsed_command}{A parsed command object}

\item{sav_info}{SAV file info from \code{\link[=parse_sav]{parse_sav()}}}
}
\value{
The parsed command with TO syntax expanded
}
\description{
Expands variable ranges like VAR1 TO VAR5 into the full list of
variables between them in the dataset.
}
\examples{
\dontrun{
parsed   <- parse_sps("analysis.sps")
sav_info <- parse_sav("data.sav")
# Expand TO syntax (e.g., item1 TO item10)
parsed <- lapply(parsed, expand_spss_variables, sav_info = sav_info)
}
}
