% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{make_vars_str}
\alias{make_vars_str}
\title{Build a c('VAR1', 'VAR2', ...) string for jmv arguments}
\usage{
make_vars_str(vars)
}
\description{
Build a c('VAR1', 'VAR2', ...) string for jmv arguments
}
