% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_converter.R
\name{normalize_spss_names}
\alias{normalize_spss_names}
\title{Normalize SPSS variable names to R-safe uppercase names}
\usage{
normalize_spss_names(x)
}
\arguments{
\item{x}{Character vector of variable names}
}
\value{
Normalized variable names
}
\description{
Normalize SPSS variable names to R-safe uppercase names
}
