% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss2rmarkdown-package.R
\docType{package}
\name{spss2rmarkdown-package}
\alias{spss2rmarkdown}
\alias{spss2rmarkdown-package}
\title{spss2rmarkdown: Convert SPSS Syntax Files to Reproducible R Markdown Reports}
\description{
Converts SPSS .sps syntax files and .sav data files into reproducible R Markdown reports. Parses 60+ SPSS commands, translates them to equivalent R code using the jmv package, and generates formatted R Markdown documents with original SPSS syntax preserved as comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giladfeldman/spss2rmarkdown}
  \item Report bugs at \url{https://github.com/giladfeldman/spss2rmarkdown/issues}
}

}
\author{
\strong{Maintainer}: Gilad Feldman \email{giladfel@gmail.com} (\href{https://orcid.org/0000-0003-2990-8997}{ORCID})

}
\keyword{internal}
